+++ apps/site/src/pages/Components/Checkbox.tsx
@@ -5,7 +5,7 @@ import {
   boolean,
   signalTarget,
   string,
-} from '@chronista-club/creoui-editor-host'
+} from '@chronista-club/creo-ui-editor-host'
 import { A } from '@solidjs/router'
 import { createSignal } from 'solid-js'
 
@@ -135,7 +135,7 @@ export default function Checkbox() {
         <div class="docs-playground-frame">
           <EditorHostProvider
             config={{
-              localStorageNamespace: 'creoui-docs.checkbox-editor',
+              localStorageNamespace: 'creo-ui-docs.checkbox-editor',
             }}
           >
             <CheckboxEditorDemo />