+++ apps/site/src/pages/Components/Radio.tsx
@@ -5,7 +5,7 @@ import {
   boolean,
   signalTarget,
   string,
-} from '@chronista-club/creoui-editor-host'
+} from '@chronista-club/creo-ui-editor-host'
 import { A } from '@solidjs/router'
 import { createSignal } from 'solid-js'
 
@@ -148,7 +148,7 @@ export default function Radio() {
         <div class="docs-playground-frame">
           <EditorHostProvider
             config={{
-              localStorageNamespace: 'creoui-docs.radio-editor',
+              localStorageNamespace: 'creo-ui-docs.radio-editor',
             }}
           >
             <RadioEditorDemo />